requirements.txt
asgiref==3.4.1
astroid==2.9.0
backports.zoneinfo==0.2.1
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dj-database-url==0.5.0
Django==4.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.2
pylint==2.12.2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.1
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.3
